feat(benchmarks): adds benchmark ci
Closes #27

Add integration-level benchmark tests and enable automatic benchmark
saving for CI/local runs so benchmark results are persisted to
directory of choice. This implements the changes requested
in issue #27 and documents the run command in CONTRIBUTING.md.
Also adds the workflow file to run and store benchmark results during CI.